Section 1 - moved to Home component dir, changed title with SectionTitle component
