+++ src/components/Home/Section_1/SectionContent/index.js
@@ -3,14 +3,13 @@ import React from 'react'
 const Section1Content = () => {
   return (
     <>
-      <div className="home-section-title title-section1">Welcome back!</div>
       <div className="home-section1-subtitle subtitle-section1">There is no substitute for in-person connections</div>
       <div className="section1-text">
         <p>At Reactathon, we faciliate conversations around the latest topics, technologies, and various skill levels to help you find the people & subjects that are most meaningful for you. Whether you are a junior engineer looking to learn from senior engineers & thought-leaders, a director of engineering seeking to push the boundaries of your own team, or just someone who loves the energy & network effect of in-person events, Reactathon has something for you.</p>
         <p>We curate a single track of talks that not only help you to learn from some of the brightest minds and best speakers in web development -- but also to facilitate more dialogue with your frontend peers. These lessons & experiences will have a direct impact not only for you and your company -- but also for your morale & overall enjoyment that will last well beyond the conference itself.</p>
       </div>
     </>
-  );
-};
+  )
+}
 
-export default Section1Content;
+export default Section1Content
