+++ src/components/Home/Section_1/index.js
@@ -1,11 +1,13 @@
 import React from 'react'
 import ViewAllSection1 from './ViewAll'
 import Section1Content from './SectionContent'
+import SectionTitle from '../SectionTitle'
 import './index.css'
 
 const Section1 = () => {
   return (
-    <section className="section-content section-section1-content">
+    <section className="section-section1-content">
+      <SectionTitle text='Welcome back!' className='title-section1' />
       <Section1Content />
       <ViewAllSection1 />
     </section>
